res made the following changes:     - Added support for buffer destination resolution by a shader       program.     - Added support for Cg program parameter names as buffer       destinations.     - Improved mapping of buffers to generic vertex attributes.     - Improved nesting of shader conditional templates a bit.     - Files included via a shader conditional <?include?> must now       have an <include> node to better comply with the XML spec.     - Fixed some VC7 warnings.     - Took SR-related projects out of the VC6 workspaces.     - Added a workaround for timing problems on Win32 on multiprocessor       machines.     - Made some reftracker-related tweas.     - Added a (incomplete and still disabled) Cg-based implementation       of the std_lighting shader.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@22787 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/movierecorder/rtjpeg.cpp
@@ -372,7 +372,7 @@ int RTjpeg_s2b(int16 *data, int8 *strm, uint8 bt8, uint32 *qtbl)
 
   switch( bitten ) {
   case 0x03:
-    data[i]= -qtbl[i];
+    data[i]= -((int32)qtbl[i]);
     break;
   case 0x02:
     goto FUSSWEG;
